CMap3DWidget:
fix bug: If do a 3D view of a high area e.g. 3000m elevation the Z axis scaling of 1 is suboptimal

1. get the min / max elevation value on the viewed area of a map
2. if dem data absent for it area, get  min / max elevation value from track
otherwice set it 0, 1
3. subtract the offset and set the Z axis scale to a somewhat sane value
in respect to the meters covered by the square edges.

The idea would be to add this basic  offset / scale factor to the OpenGL
transformation matrix and to use the Z axis zoom.

thanks to Oliver for this idea and maps for experiments.

in additional: zZoom work faster now.


git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@854 d0c010d8-bb82-477e-a075-112140230451